# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/volumes": {
      "get": {
        "tags": [
          "volume"
        ],
        "summary": "볼륨 목록 조회",
        "operationId": "find_volumes_detail_volumes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/VolumeSortOption",
              "default": "created_at"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumesDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 정보가 유효하지 않은 경우"
          },
          "422": {
            "description": "요청 데이터의 값이나 형식이 잘못된 경우"
          }
        }
      },
      "post": {
        "tags": [
          "volume"
        ],
        "summary": "볼륨 생성",
        "operationId": "create_volume_volumes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVolumeRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumeResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 정보가 유효하지 않은 경우"
          },
          "422": {
            "description": "요청 데이터의 값이나 형식이 잘못된 경우"
          }
        }
      }
    },
    "/volumes/{volume_id}": {
      "get": {
        "tags": [
          "volume"
        ],
        "summary": "볼륨 단건 조회",
        "operationId": "get_volume_detail_volumes__volume_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Volume Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumeDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 정보가 유효하지 않은 경우"
          },
          "403": {
            "description": "볼륨에 대한 접근 권한이 없는 경우"
          },
          "404": {
            "description": "볼륨을 찾을 수 없는 경우"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "volume"
        ],
        "summary": "볼륨 삭제",
        "operationId": "delete_volume_volumes__volume_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Volume Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "인증 정보가 유효하지 않은 경우"
          },
          "403": {
            "description": "볼륨에 대한 접근 권한이 없는 경우"
          },
          "404": {
            "description": "볼륨을 찾을 수 없는 경우"
          },
          "409": {
            "description": "볼륨에 연결된 서버가 존재하거나, 볼륨이 삭제 가능한 상태가 아닌 경우"
          },
          "422": {
            "description": "요청 데이터의 값이나 형식이 잘못된 경우"
          }
        }
      }
    },
    "/volumes/{volume_id}/info": {
      "put": {
        "tags": [
          "volume"
        ],
        "summary": "볼륨 정보 변경",
        "operationId": "update_volume_info_volumes__volume_id__info_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Volume Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVolumeInfoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumeResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 정보가 유효하지 않은 경우"
          },
          "403": {
            "description": "볼륨에 대한 접근 권한이 없는 경우"
          },
          "404": {
            "description": "볼륨을 찾을 수 없는 경우"
          },
          "409": {
            "description": "변경하려는 이름이 이미 사용중인 경우"
          },
          "422": {
            "description": "요청 데이터의 값이나 형식이 잘못된 경우"
          }
        }
      }
    },
    "/volumes/{volume_id}/size": {
      "put": {
        "tags": [
          "volume"
        ],
        "summary": "볼륨 용량 변경",
        "operationId": "update_volume_size_volumes__volume_id__size_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Volume Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVolumeSizeRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumeResponse"
                }
              }
            }
          },
          "400": {
            "description": "변경하려는 볼륨 용량이 기존 용량보다 크지 않은 경우"
          },
          "401": {
            "description": "인증 정보가 유효하지 않은 경우"
          },
          "403": {
            "description": "볼륨에 대한 접근 권한이 없는 경우"
          },
          "404": {
            "description": "볼륨을 찾을 수 없는 경우"
          },
          "422": {
            "description": "요청 데이터의 값이나 형식이 잘못된 경우"
          }
        }
      }
    },
...
    "/floating-ips": {
      "get": {
        "tags": [
          "floating-ip"
        ],
        "summary": "소유한 플로팅 IP 목록 조회",
        "operationId": "find_floating_ips_floating_ips_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FloatingIpSortOption",
              "default": "created_at"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloatingIpDetailsResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 정보가 유효하지 않은 경우"
          },
          "422": {
            "description": "쿼리 파라미터 값이나 형식이 잘못된 경우"
          }
        }
      },
      "post": {
        "tags": [
          "floating-ip"
        ],
        "summary": "플로팅 IP 할당",
        "operationId": "create_floating_ip_floating_ips_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFloatingIpRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloatingIpResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 정보가 유효하지 않은 경우"
          },
          "422": {
            "description": "요청 데이터의 값이나 형식이 잘못된 경우"
          }
        }
      }
    },
    "/floating-ips/{floating_ip_id}": {
      "get": {
        "tags": [
          "floating-ip"
        ],
        "summary": "단일 플로팅 IP 조회",
        "operationId": "get_floating_ip_floating_ips__floating_ip_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "floating_ip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Floating Ip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloatingIpDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 정보가 유효하지 않은 경우"
          },
          "403": {
            "description": "floating ip에 대한 접근 권한이 없는 경우"
          },
          "404": {
            "description": "해당 ID의 플로팅 IP를 찾을 수 없는 경우"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "floating-ip"
        ],
        "summary": "플로팅 IP 할당 해제(삭제)",
        "operationId": "delete_floating_ip_floating_ips__floating_ip_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "floating_ip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Floating Ip Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "인증 정보가 유효하지 않은 경우"
          },
          "403": {
            "description": "floating ip에 대한 접근 권한이 없는 경우"
          },
          "404": {
            "description": "해당 ID의 플로팅 IP를 찾을 수 없는 경우"
          },
          "409": {
            "description": "서버에 연결된 상태에서는 삭제할 수 없음"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateFloatingIpRequest": {
        "properties": {
          "floating_network_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 36,
            "title": "Floating Network Id",
            "description": "플로팅 IP를 할당받을 Public network ID",
            "examples": [
              "b1321e7e-7709-47d1-b851-5e6c4e0e1111"
            ]
          }
        },
        "type": "object",
        "required": [
          "floating_network_id"
        ],
        "title": "CreateFloatingIpRequest"
      },
...
      "CreateVolumeRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name",
            "description": "Name of volume",
            "examples": [
              "volume-001"
            ]
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "title": "Description",
            "description": "Description of volume",
            "examples": [
              "For test"
            ]
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "description": "용량(GiB)",
            "examples": [
              1
            ]
          },
          "type_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 36,
            "title": "Type Id",
            "description": "사용할 volume type의 uuid",
            "examples": [
              "64abcd22-a30b-4982-8f82-332e89ff4bf7"
            ]
          },
          "image_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 36,
            "title": "Image Id",
            "description": "사용할 부팅 이미지의 uuid",
            "examples": [
              "1abc7a2f-8eec-49a2-b9ef-be16e0959cdb"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "size",
          "type_id",
          "image_id"
        ],
        "title": "CreateVolumeRequest"
      },
...
      "FloatingIpDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "플로팅 IP ID",
            "examples": [
              1
            ]
          },
          "address": {
            "type": "string",
            "title": "Address",
            "description": "IP 주소",
            "examples": [
              "0.0.0.0"
            ]
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "프로젝트 ID",
            "examples": [
              1
            ]
          },
          "status": {
            "$ref": "#/components/schemas/FloatingIpStatus",
            "description": "플로팅 IP 상태",
            "examples": [
              "ACTIVE"
            ]
          },
          "server": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/router__floating_ip__response__ServerResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "연결된 서버 정보"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "생성일"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "수정일"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At",
            "description": "삭제일"
          }
        },
        "type": "object",
        "required": [
          "id",
          "address",
          "project_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "FloatingIpDetailResponse"
      },
      "FloatingIpDetailsResponse": {
        "properties": {
          "floating_ips": {
            "items": {
              "$ref": "#/components/schemas/FloatingIpDetailResponse"
            },
            "type": "array",
            "title": "Floating Ips"
          }
        },
        "type": "object",
        "required": [
          "floating_ips"
        ],
        "title": "FloatingIpDetailsResponse"
      },
      "FloatingIpResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "플로팅 IP ID",
            "examples": [
              1
            ]
          },
          "address": {
            "type": "string",
            "title": "Address",
            "description": "IP 주소",
            "examples": [
              "0.0.0.0"
            ]
          },
          "server_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Id",
            "description": "서버 ID",
            "examples": [
              1
            ]
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "프로젝트 ID",
            "examples": [
              1
            ]
          },
          "status": {
            "$ref": "#/components/schemas/FloatingIpStatus",
            "description": "플로팅 IP 상태",
            "examples": [
              "ACTIVE"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "생성일"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "수정일"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At",
            "description": "삭제일"
          }
        },
        "type": "object",
        "required": [
          "id",
          "address",
          "project_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "FloatingIpResponse"
      },
      "FloatingIpSortOption": {
        "type": "string",
        "enum": [
          "address",
          "created_at"
        ],
        "title": "FloatingIpSortOption"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder"
      },
...
      "UpdateVolumeInfoRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name",
            "description": "Name of volume",
            "examples": [
              "volume-001"
            ]
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "title": "Description",
            "description": "Description of volume",
            "examples": [
              "For test"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "description"
        ],
        "title": "UpdateVolumeInfoRequest"
      },
      "UpdateVolumeSizeRequest": {
        "properties": {
          "size": {
            "type": "integer",
            "title": "Size",
            "description": "변경하려는 용량(GiB). 반드시 기존 용량보다 커야 합니다.",
            "examples": [
              4
            ]
          }
        },
        "type": "object",
        "required": [
          "size"
        ],
        "title": "UpdateVolumeSizeRequest"
      },
...
      "VolumeDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Id of volume"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "Id of project"
          },
          "server": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/router__volume__response__ServerResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Server information"
          },
          "type_id": {
            "type": "string",
            "title": "Type Id",
            "description": "Id of volume type",
            "examples": [
              "64abcd22-a30b-4982-8f82-332e89ff4bf1"
            ]
          },
          "image_id": {
            "type": "string",
            "title": "Image Id",
            "description": "Id of boot image",
            "examples": [
              "2fcc7a2f-8eec-49a2-b9ef-be16e0959cdb"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Name of volume",
            "examples": [
              "volume-001"
            ]
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Description of volume",
            "examples": [
              "volume-001"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/VolumeStatus",
            "description": "Status of volume"
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "description": "Size of volume(GiB)"
          },
          "is_root_volume": {
            "type": "boolean",
            "title": "Is Root Volume",
            "description": "Whether it is a root volume"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last update time"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At",
            "description": "Deletion time"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "type_id",
          "image_id",
          "name",
          "description",
          "status",
          "size",
          "is_root_volume",
          "created_at",
          "updated_at"
        ],
        "title": "VolumeDetailResponse"
      },
      "VolumeResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Id of volume"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "Id of project"
          },
          "server_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Id",
            "description": "Id of server"
          },
          "type_id": {
            "type": "string",
            "title": "Type Id",
            "description": "Id of volume type",
            "examples": [
              "64abcd22-a30b-4982-8f82-332e89ff4bf1"
            ]
          },
          "image_id": {
            "type": "string",
            "title": "Image Id",
            "description": "Id of boot image",
            "examples": [
              "2fcc7a2f-8eec-49a2-b9ef-be16e0959cdb"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Name of volume",
            "examples": [
              "volume-001"
            ]
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Description of volume",
            "examples": [
              "volume-001"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/VolumeStatus",
            "description": "Status of volume"
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "description": "Size of volume(GiB)"
          },
          "is_root_volume": {
            "type": "boolean",
            "title": "Is Root Volume",
            "description": "Whether it is a root volume"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last update time"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At",
            "description": "Deletion time"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "type_id",
          "image_id",
          "name",
          "description",
          "status",
          "size",
          "is_root_volume",
          "created_at",
          "updated_at"
        ],
        "title": "VolumeResponse"
      },
      "VolumeSortOption": {
        "type": "string",
        "enum": [
          "created_at",
          "name"
        ],
        "title": "VolumeSortOption"
      },
      "VolumeStatus": {
        "type": "string",
        "enum": [
          "CREATING",
          "AVAILABLE",
          "RESERVED",
          "ATTACHING",
          "DETACHING",
          "IN_USE",
          "MAINTENANCE",
          "DELETING",
          "AWAITING_TRANSFER",
          "ERROR",
          "ERROR_DELETING",
          "BACKING_UP",
          "RESTORING_BACKUP",
          "ERROR_BACKING_UP",
          "ERROR_RESTORING",
          "ERROR_EXTENDING",
          "DOWNLOADING",
          "UPLOADING",
          "RETYPING",
          "EXTENDING"
        ],
        "title": "VolumeStatus"
      },
      "VolumesDetailResponse": {
        "properties": {
          "volumes": {
            "items": {
              "$ref": "#/components/schemas/VolumeDetailResponse"
            },
            "type": "array",
            "title": "Volumes"
          }
        },
        "type": "object",
        "required": [
          "volumes"
        ],
        "title": "VolumesDetailResponse"
      },
...
      "router__volume__response__ServerResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Id of server"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Name of server",
            "examples": [
              "server-001"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last update time"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At",
            "description": "Deletion time"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "title": "ServerResponse"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}